Component Props The component takes in two props:
submission: This is an object that contains form submission data, specifically a content field which holds a mapping of question labels to answers.

index: The index of the current submission, used for numbering the response (e.g., "Response - 1", "Response - 2", etc.).

These props allow the component to render one submission from a list of many.