Conditional Rendering The component conditionally renders two different UIs depending on whether the user is subscribed or not:
If Subscribed
Displays a short message stating the user has a subscription and can create more forms freely.

No progress bar or upgrade suggestion is shown.

If Not Subscribed
Displays a progress bar indicating how many of the free forms have been used.

Displays a message with a counter (e.g., "2 out of 3 forms generated").

Includes a link to the upgrade page, encouraging the user to subscribe to generate more forms.

This offers clear, real-time feedback to the user and a seamless way to upgrade when necessary.