const user = await currentUser();:
This line fetches the currently authenticated user using Clerk's currentUser method.

It is asynchronous because currentUser likely makes a request to Clerk’s server to fetch the current session’s user details.

The result of await currentUser() is stored in the user variable. If a user is logged in, it contains their user data (like id, email, etc.). If no user is logged in, this variable will be null.